Move all of the DevTools commands to a new cog
This update has also moved all of the DevTools commands to a new slash command group, to make finding the commands easier.

(only developers will understand 🤫)
