+++ main.py
@@ -620,6 +620,7 @@ active_cogs = (
     "fun",
     "utils",
     "afk",
+    "devtools",
     # "osu",  Disabled due to ossapi library metadata issues. (will probably remove osu cog anyway, because cog code is outdated with ossapi library)
     "weather",
     "isocard"
